pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["liblaf <i@liblaf.me>"]
description = "Repository Template"
license = "MIT"
name = "template"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
pydantic = "^2.6.1"
python = "^3.11"
pyyaml = "^6.0.1"
rich = "^13.7.0"
typer = "^0.9.0"
types-pyyaml = "^6.0.12.12"

[tool.poetry.group.dev.dependencies]
poetry = "^1.7.1"
ruff = "^0.2.1"

[tool.poetry.scripts]
template = "template.cli:app"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "SIM", "UP"]
